build: [ANDROAPP-5372] Merge TableItem and TableList (#3202)
* build: [ANDROAPP-5372] Merge TableItem and TableList
